Move random string function to utils
